+++ app/src/main/java/com/everest/movieapp/ui/main/PlaceholderFragment.kt
@@ -8,8 +8,8 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import com.everest.movieapp.R
 import com.everest.movieapp.databinding.FragmentMainBinding
+import com.everest.movieapp.viewmodel.PageViewModel
 
 /**
  * A placeholder fragment containing a simple view.
